% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{make.stability}
\alias{make.stability}
\title{making network stability}
\usage{
make.stability(val.mat, biomasses, losses, efficiencies, growth.rate,
  losses.scale = NULL, bioms.prefs = TRUE, bioms.losses = TRUE,
  ef.level = "prey", interval = c(1e-12, 1), ...)
}
\arguments{
\item{val.mat}{A matrix describing fluxes between species (usually a result of \code{\link[fluxweb]{fluxing}} function).}

\item{biomasses}{A vector of species biomasses.}

\item{losses}{A vector or an array of species energy losses (excluding predation).}

\item{efficiencies}{A vector or an array of conversion efficiencies of species in the adjacency matrix. These values describe the proportion of consumed energy that is converted to biomass of the consumer.}

\item{growth.rate}{A vector defining growth rate of basal species.}

\item{losses.scale}{Defines a Column from \code{losses} mulitplicator value will apply to. (default \code{NULL} if multiplicator independant of losses).}

\item{bioms.prefs}{Logical, if \code{TRUE} (default) preferences are scaled accordingly to species biomasses.}

\item{bioms.losses}{Logical, if \code{TRUE} (default) losses are scaled with biomass.}

\item{ef.level}{Set to \code{"prey"} if efficiences are defined by prey, \code{"pred"} if they are a property of the predator.}

\item{interval}{Search interval for returned value.}

\item{...}{Optional parameters for function \code{\link[stats]{uniroot}}}
}
\value{
A list from \code{\link[stats]{uniroot}} function.
}
\description{
Find the smallest multiplicator of a variable from losses insuring system stability - !Version under devlopment!
}
\details{
The function assume a monotonous increase of stability with multiplicator value. Solution is estimated from the \code{\link[stats]{uniroot}} function, and stability using the \code{\link[fluxweb]{fluxing}} function
Thus, accordingly to \code{\link[stats]{uniroot}} solving criteria, if stability values at the two extremu parts of the interval are of same sign, an error is raised.

Behavior of the multiplicative term depends on the type of losses:
\itemize{
\item{\code{losses.scale = NULL} and \code{is.vector(losses)}:} multiplicator will be applied to the \code{losses} vector.
\item{\code{losses.scale = NULL} and \code{is.matrix(losses)}:} multiplicator will be independant of any columns from \code{losses}.
\item{\code{losses.scale = FALSE} :} multiplicator always independant of losses.
\item{other values:} should refer to an element of losses.

This is an ugly part, but I don't see how to do better right now. Any ideas?

}
}
\examples{

mat = matrix(0, 5, 5)
mat[2,1] = 1
mat[5,1] = 1
mat[3,2] = 1
mat[4,2] = 1

# metabolic losses:
losses = c(0.6299, 0.25, 0.12, 0.15, 0.1)

# biomasses
biomasses = c(1.5, 1.51, 1.8, 1.5, 2)

# efficiencies
efficiencies = c(0.85, 0.85, 0.35, 0.35, 0.35)
# growth rates of basal sppecies
growth.rates = c(0.28, 0.32, 0.19)
losses = c(0.64, 0.4, 0.25, 0.3, 0.1)
val.mat = fluxing(mat, biomasses, losses, efficiencies, bioms.pref = TRUE, ef.level = "pred")
make.stability(val.mat, biomasses, losses, efficiencies, growth.rates, ef.level = "pred")


}
\seealso{
\code{\link[stats]{uniroot}} for root estimate and \code{\link[fluxweb]{stability.value}} for assessing system stability.
}
