% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.level.R
\name{species.level}
\alias{species.level}
\title{Food web of a soil network ecosystem and species general information.}
\format{a list of 5 elements:
\describe{
 \item{mat}{the network adjacency matrix}
 \item{biomasses}{species biomasses (g)}
 \item{bodymasses}{species bodymasses (g)}
 \item{efficiencies}{species assimilation efficiencies}
 \item{names}{species names}
}}
\description{
This dataset contains the matrix describing trophic interactions from a deutsch soil food-web (Digel et al. 2014, Oikos)
as well as some ecological information on species: biomasses, body masses and and species names.
}
